# Create and add instance reference inside ResolveMetadata*