--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Object/ObjectDefaultConverter.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/Converters/Object/ObjectDefaultConverter.cs
@@ -77,7 +77,7 @@ namespace System.Text.Json.Serialization.Converters
                 {
                     if (options.ReferenceHandler != null)
                     {
-                        if (JsonSerializer.ResolveMetadata(this, ref reader, ref state))
+                        if (JsonSerializer.ResolveMetadataForJsonObject(this, ref reader, ref state, options))
                         {
                             if (state.Current.ObjectState == StackFrameObjectState.ReadRefEndObject)
                             {
@@ -91,8 +91,6 @@ namespace System.Text.Json.Serialization.Converters
                             return false;
                         }
                     }
-
-                    state.Current.ObjectState = StackFrameObjectState.PropertyValue;
                 }
 
                 if (state.Current.ObjectState < StackFrameObjectState.CreatedObject)
@@ -103,13 +101,6 @@ namespace System.Text.Json.Serialization.Converters
                     }
 
                     obj = state.Current.JsonClassInfo.CreateObject!()!;
-                    if (state.Current.MetadataId != null)
-                    {
-                        state.ReferenceResolver.AddReference(state.Current.MetadataId, obj);
-                        // Clear metadata name, if the next read fails
-                        // we want to point the JSON path to the property's object.
-                        state.Current.JsonPropertyName = null;
-                    }
 
                     state.Current.ReturnValue = obj;
                     state.Current.ObjectState = StackFrameObjectState.CreatedObject;
@@ -415,6 +406,17 @@ namespace System.Text.Json.Serialization.Converters
             }
 
             return true;
+        }
+
+        public sealed override void CreateInstance(ref Utf8JsonReader reader, ref ReadStack state, JsonSerializerOptions options)
+        {
+            if (state.Current.JsonClassInfo.CreateObject == null)
+            {
+                ThrowHelper.ThrowNotSupportedException_DeserializeNoConstructor(state.Current.JsonClassInfo.Type, ref reader, ref state);
+            }
+
+            object obj = state.Current.JsonClassInfo.CreateObject!()!;
+            state.Current.ReturnValue = obj;
         }
     }
 }