--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonResumableConverterOfT.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/JsonResumableConverterOfT.cs
@@ -41,7 +41,5 @@ namespace System.Text.Json.Serialization
         }
 
         public sealed override bool HandleNull => false;
-
-        public abstract void CreateInstance(ref Utf8JsonReader reader, ref ReadStack state, JsonSerializerOptions options);
     }
 }