--- a/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReadStackFrame.cs
+++ b/src/libraries/System.Text.Json/src/System/Text/Json/Serialization/ReadStackFrame.cs
@@ -32,8 +32,8 @@ namespace System.Text.Json
         public JsonClassInfo JsonClassInfo;
         public StackFrameObjectState ObjectState; // State tracking the current object.
 
-        // Preserve reference.
-        public string? MetadataId;
+        // Validate EndObject token on array with preserve semantics.
+        public bool ValidateEndTokenOnArray;
 
         // For performance, we order the properties by the first deserialize and PropertyIndex helps find the right slot quicker.
         public int PropertyIndex;
@@ -59,7 +59,7 @@ namespace System.Text.Json
             JsonPropertyName = null;
             JsonPropertyNameAsString = null;
             PropertyState = StackFramePropertyState.None;
-            MetadataId = null;
+            ValidateEndTokenOnArray = false;
 
             // No need to clear these since they are overwritten each time:
             //  UseExtensionProperty