...in order to remove workaround described in https://github.com/dotnet/runtime/issues/37168 used to correctly build the JSON path in case of error.

This PR adds a new method `CreateInstance` to `JsonResumableConverter<T>` that is used to instantiate the object at the time that `ReferenceResolver.AddReference` is called.

It also splits `ResolveMetadata` into two methods: `ResolveMetadataForJsonObject` (Used on POCOs and dictionaries) and `ResolveMetadataForJsonArray` (Used for collections such as `List<T>`).

This makes the code more mantainable, reduces the amount of if checks and allow us to remove the string field `ReadStackFrame.MetadataId` from the state.

Fixes https://github.com/dotnet/runtime/issues/37168.